+++ Gwen/Dto/Riot/LolChallenges/ChallengeInfo.cs
@@ -2,10 +2,10 @@
 {
     public record ChallengeInfo
     {
-        public int ChallengeId { get; init; }
-        public int Percentile { get; init; }
+        public long ChallengeId { get; init; }
+        public double Percentile { get; init; }
         public string Level { get; init; } = default!;
-        public int Value { get; init; }
+        public double Value { get; init; }
         public long AchievedTime { get; init; }
     }
 }
